Build out categories as a select drop down
There is no apparent order to the categories, other than coming
over from a spreadsheet laid out as so. Eating out is a bad habit
and happens to drop into the first column. For shame.

Donations is last after previously being a separate worksheet
entirely, but started rolling it into the normal flow in order
to have some lookups work easier. Moot given this project.